# chore: general performance optimisations