--- a/.changeset/calm-cherries-lay.md
+++ b/.changeset/calm-cherries-lay.md
@@ -1,0 +1,5 @@
+---
+"preact-render-to-string": patch
+---
+
+General performance optimisations