--- a/benchmarks/index.js
+++ b/benchmarks/index.js
@@ -4,7 +4,7 @@ import renderToStringBaseline from 'baseline-rts';
 // import renderToString from '../src/index';
 import renderToString from '../dist/index.module.js';
 import TextApp from './text';
-// import StackApp from './stack';
+import StackApp from './stack';
 import { App as IsomorphicSearchResults } from './isomorphic-ui/search-results/index';
 import { App as ColorPicker } from './isomorphic-ui/color-picker';
 
@@ -19,6 +19,5 @@ function suite(name, Root) {
 	await suite('Text', TextApp);
 	await suite('SearchResults', IsomorphicSearchResults);
 	await suite('ColorPicker', ColorPicker);
-	// TODO: Enable this once we switched away from recursion
-	// await suite('Stack Depth', StackApp);
+	await suite('Stack Depth', StackApp);
 })();