--- a/package.json
+++ b/package.json
@@ -132,7 +132,7 @@
 		"@babel/register": "^7.12.10",
 		"@changesets/changelog-github": "^0.4.1",
 		"@changesets/cli": "^2.18.0",
-		"baseline-rts": "npm:preact-render-to-string@latest",
+		"baseline-rts": "npm:preact-render-to-string@6.5.7",
 		"benchmarkjs-pretty": "^2.0.1",
 		"chai": "^4.2.0",
 		"check-export-map": "^1.3.1",